fix(records): give the link reads and alarm ladders their C owners
db_get_link took over setLinkAlarm, one deferred dbGetLink owner replaced
the per-record multi-input loops, and recGblGetTimeStamp restamps a seq
link group once. The analog limits now live and compare in the record's
own DBF, which is what lets int64 HYST reach the ladder at all.

32 per-finding commits: caucus/VS754JQF2H/fix-rec-098446bf-1

+++ crates/epics-base-rs/src/server/records/int64in.rs
@@ -50,10 +50,12 @@ pub struct Int64inRecord {
     #[field(type = "Int64")]
     pub hyst: i64,
     // LALM is DBF_INT64 too, but `special(SPC_NOMOD)` (dbd:233-236): read-only,
-    // so no client put reaches the parse. Kept f64 — it is the internal
-    // last-alarmed bookkeeping the alarm filter reads as a double.
-    #[field(type = "Double")]
-    pub lalm: f64,
+    // so no client put reaches the parse. It is the alarm ladder's own
+    // `epicsInt64 lalm` (`int64inRecord.c:262`) — the value the hysteresis
+    // arm compares a threshold against — so it holds that threshold exactly
+    // rather than as a rounded double.
+    #[field(type = "Int64")]
+    pub lalm: i64,
     #[field(type = "Int64")]
     pub adel: i64,
     #[field(type = "Int64")]
@@ -101,7 +103,7 @@ impl Default for Int64inRecord {
             hopr: 0,
             lopr: 0,
             hyst: 0,
-            lalm: 0.0,
+            lalm: 0,
             adel: 0,
             mdel: 0,
             aftc: 0.0,
